net, ipv4, tcp_noord: fix unit of measurement for tx_timer-related variables
The tx_timer variable of the noord structure is, at the moment, unused;
however, it should be used (and perhaps will be used) to set the RTO
according to TCP Noordwijk's needs. The RTO timeout is stored in
eights of microseconds: this commit prepares the variables to be
used for that purpose.

Signed-off-by: Daniele Cristofori <daniele.cristofori89@gmail.com>
Signed-off-by: Stefano Alletto <alletto.s@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>