net, ipv4, tcp_noord: improve estimation of burst duration
Currently, the duration of a burst is estimated as the difference
between the timestamps of the last and first ACKs of the burst.
However, mechanisms such as "delayed ACK" might reduce that temporal
distance and force Noordwijk to underestimate the duration of the
burst. This commit lets Noordwijk use the TX_START event and the
ACK count to estimate the duration of the burst. Please see the
code for further details.

Signed-off-by: Daniele Cristofori <daniele.cristofori89@gmail.com>
Signed-off-by: Stefano Alletto <alletto.s@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>