Revert "tcp: allow drivers to tweak TSQ logic"
There are a lot of drivers that are still outside the mainline, and it is
impossible to expect the authors (which in most cases do not maintain the
driver anymore) to check the best value for the device, and then to insert
it into the driver code. Moreover, if the users want to check or to use
other values, they have to recompile the driver (or the kernel).

This reverts commit 3a9b76fd0db9f0d426533f96a68a62a58753a51e, preparing the
field for tunable sysctl knobs.

Signed-off-by: Natale Patriciello <natale.patriciello@gmail.com>
Signed-off-by: Carlo Augusto Grazia <carloaugusto.grazia@unimore.it>